requirements.txt
datasets
chromadb
open-clip-torch
google-generativeai
streamlit